Add support for mobi7 only and print replica (pdf)

+++ tests/test_mobi.py
@@ -1,5 +1,6 @@
+import os
 import shutil
-from os.path import abspath, join, exists, dirname
+from os.path import abspath, join, exists, dirname, splitext
 import mobi
 
 
@@ -7,7 +8,10 @@ TEST_DIR = dirname(abspath(__file__))
 
 
 def test_extract():
-    tempdir, epub_filepath = mobi.extract(join(TEST_DIR, "demo.mobi"))
-    assert exists(tempdir)
-    assert exists(epub_filepath)
-    shutil.rmtree(tempdir)
+    for fname in os.listdir(TEST_DIR):
+        ext = splitext(fname)[-1].upper()
+        if ext in [".MOBI", ".PRC", ".AZW", ".AZW3", ".AZW4"]:
+            tempdir, epub_filepath = mobi.extract(join(TEST_DIR, fname))
+            assert exists(tempdir)
+            assert exists(epub_filepath)
+            shutil.rmtree(tempdir)
